UPBGE: Fix NDOF input event issues.
Previously pressing or releasing any keys was exiting the game engine.
The issue was the cause of the WITH_NDOF_INPUT macro not set in
the parts using GHOST when the user compiled with NDOF.
Without this macro all the values was moveed after GHOST_kEventNDOFMotion
and GHOST_kEventNDOFButton in GHOST_Types.h.
This move replace the GHOST_kEventKeyDown sended by the GHOST_kEventQuit
handled.

To fix the problem, the macro WITH_NDOF_INPUT is added in the Device/
part when it is wanted.

In the same time the wrong fix 19f48ee2271868731c24621e7499833314dc6e3f
is reverted.