UPBGE: Securise std::vector erase iteration.
Using a fixed end iterator or a incrementing the iterator after
delete its previous is unsafe for std::vector. Because erasing
an element can realloc the whole container and then invaldate
the existing iterators.

To fix this issue we compare with the current returned of .end()
every frame and assign the iterator to the returned value of
.erase instead of incrementing the iterator.