UPBGE: Add refraction option for cube maps.
Recently reflections and in particular realtime reflections were added
in the UPBGE, but refractions were lacking to do realistic transparent
shapes.

To erase this lack, fake refractions are added. They are fake because
they depends only in the the first saw face and not the total faces
traversed by the light, which in some conditions are very crucial.
For example a sphere refractions are inverted and a plane refractions
are not. The refractions implemented don't invert the light.
These refractions use as argument an Indice Of Refraction (IOR) and a
Refraction Ratio value of refractions mixed with reflections.

From the user point of view, the UI changed and the python API too.

First the UI now takes two values exposed in the Influences panel in the
Texture panel. These two values are named "IOR" and "Ratio"
for the refraction ratio.

In the python api similar changed are made: two attributes are added to
BL_Texture:

BL_Texture.ior
BL_Texture.refractionRatio

The values have a similar name as the UI and as the same effect.

From the internal point of view, the material shader was modified to support
refractions. But first as we use a ratio of refraction mixed with reflection
a new function was added to do it directly in the shader:

mtex_cube_map_refl_refr

This function depends on two other functions which compute respectively
reflection and refraction:

mtex_cube_map_refl_color
mtex_cube_map_refr_color

If the ratio value passed to the GLSL function mtex_cube_map_refl_refr is
to one, then only the refraction color is computed, else if the ratio is to
zero only the reflection color is computed else the refraction and the
reflection color are computed and mixed with the given ratio value.