UPBGE: Implement HDR for ImageRender and ImageMirror.
Previously the user was able to set Image[Render/Mirror]
multisample but not image format (or called HDR).

To solve this issue the python constructor of the both
class take as last argument the HDR value, which is one
of the new value added in bge.render:
bge.render.HDR_NONE
bge.render.HDR_HALF_FLOAT
bge.render.HDR_FULL_FLOAT

Internally the image format is converted to be used by
off screens and the format value is passed to the function
loadTexture.