UPBGE: Implement material switch using KX_MeshProxy.replaceMaterial(index, material).
Previously the user wasn't able to replace the material of a mesh.
A workaround was to use the same meshes a lot of time but with
different materials. This workaround use a huge extra memory.

To solve this issue the user is now able to replace one material
of a mesh by indicating its slot index in the mesh and a pointer
to the new material to use in the python function:
KX_MeshProxy.replaceMaterial(index, material)

But the replacement can failed if:
- the new material doesn't have the same scene than the previous
 material.
- the index is invalid

Internally the display array bucket representing the couple
mesh and previous material is moved to the new material and all
the mesh slots reset theirs material bucket pointer.
In the same time the attribute's layers which depends on the
combinaison of mesh and material are regenerated.