UPBGE: Begin to work on planar reflections
I just copied the cubemap structure for planar reflections and made it
compilable.

I think we have all what we need in ImageMirror to do the stuff. What is
in image mirror constructor (at the end of ImageRender.cpp) could be moved to ? and what is in
ImageRender::Render() about mirrors could be moved to ?.

As TwisterGE said we'll need 2 textures (1 for reflection and 1 for
refraction) contrarly to cubemaps.

If I say that is that because for cubemap, we needed an unique cubemap
texture to do both reflection and refraction -> we just had, with the same
uniform samplerCube, to change mapping coordinates inside the shader.

With planar reflections, we need 2 different images. If for example the
plane is horizontal, we need an image of what is upside the plane and
another image that is the capture of what is downside the plane.

So we need 2 textures (2 texture slots, 2 shaders...). As in Water Demo from martinsh.

I made an UI in Image or Movie.

We'll also need 2 new shaders in gpu_material.c below "mtex_image" ->
if tex->planarflag == TEX_PLANAR_REFLECTION: "mtex_image_refl" and if
tex->planarflag == TEX_PLANAR_REFRACTION: "mtex_image_refr"
The first shader would take as arguments sampler2D reflectedImage. The
second would take as argument sampler2D refractedImage.

The blending beeween the 2 textures would be done maybe with blend type:
mix/multiply or influence: color  ->   to adjust amount of each texture.

I'm not sure about all what I say but this is as I see the stuff for the
moment.