UPBGE: Fix ImageViewport.
Since the commit introducing the off screens render, the
ImageRender was using a constructor without arguments which
initialize nothing. But this constructor is called when in
python we do:

texture.source = ImageViewport()

To solve this issue the empty constructor is filled with
instructions similar to the second constructor. The only
problem is to get the viewport, the viewport is get from
the global canvas viewport (not canvas area) because the
ImageViewport made a copy of the screen and not a copy of an
off-screen.