UPBGE: Fix realtime cube map image free.
Previously when an image used by a realtime cube map was freed,
if the user undo some scene modifications the image became black
and raise a crash at the next game run because it become totally
invalid.

This issue was caused by the fact that the image uses a frame cache
which is used during the undo. But this cache was invalid after
calling GPU_free_image.

To solve this issue, the call to GPU_free_image is replaced by a
call to BKE_image_free_buffers. The function BKE_image_free_buffers
call internally GPU_free_image, so the image is still freed, but
before free the image, the frame cache and some other datas are
freed.