UPBGE: Implement world exposure and range in python.
Previously the world exposure and range were not exposed in python
and editable (used under dynamics uniforms).

To solve this lack, the linear factor and logarithm factor depending
on the world exposure and range are now set as uniform, but theirs
values are stored in GPUWorld and set by the function
GPU_update_exposure_range.

The exposure and range are now expossed in python as attributes names
KX_WorldInfo.exposure [0; 1]
KX_WorldInfo.range [0.2; 5]