UPBGE: Pass rendered camera to scene.pre_draw/pre_draw_setup callbacks.
Previously the user was able to altera	te rendering datas through
the callbacks scene.pre_draw/post_draw/pre_draw_setup. But he
wasn't able to know which camera was used when this callbacks
was called in case of multiple viewports.
Getting this info can help on setting an UI with a more modular
camera viewport management.

To solve this issue the scenes callbacks, exepted post_draw because
this one is not per camera, take as optional argument the current
rendered camera.

Internally the function KX_Scene::RunDrawingCallbacks takes
as first argument an enumeration coresponding to the callbacks list
instead of the callbacks list previously. This modification allow
to remove the two getter to the python callbacks list from
KX_Scene.