UPBGE: Implement text dimensions in python API.
Previously the user wasn't able to acces to the text dimensions
of a text/font object. This can be useful to make a complex UI
system.

To remove this lack, a python attributes named
KX_FontObject.dimensions is added. It returns the text
dimensions in blender units using the BLF function
BLF_width_and_height.