UPBGE: Remove mathutils error when writing to obj.worldScale.
An error saying that obj.worldScale is read only was raise when
modifying the world scale, but this world scale can be direclty
set by using = to the whole vector.

This error is removed as the assignement to the vector was allowed
and that none test showed an error with parenting and world scale.