UPBGE: Forbid setParent to an object in an inactive layer
Previously object was able to be parented between active
and inactive objects. But this was introducing big troubles
when adding an inactive object with active object children.

To solve this issue the inter layer parenting is disallowed.
Only active objects can be parented to active objects and
the same for inactive objects.