Try to implement sky texture option in environment light settings
The world texture color is applied to all objects in the scene.
We can set the intensity of this color with energy option.

Note: Maybe a second lodbias option could be added in World struct to have
a different lodbias for world texture and for the color picked up on the
texture. A lodbias option in environment light settings.