UPBGE: Reduce KX_MeshProxy creation for KX_PolyProxy.
Previously each time a KX_PolyProxy was created when the user
acces to the polygon of a mesh in python, a new KX_MeshProxy
was also possibly created when the user did:

polygon.getMesh()

To avoid duplicating the mesh proxies, we follow the behaviour
of KX_VertexProxy: Pass the mesh proxy in the constructor
instead of the RAS_MeshObject directly. After this the mesh
proxy is stored in the KX_PolyProxy instance.

By this modification the number of maximal mesh proxy is
reduced and the mesh proxy is available from the KX_PolyProxy.
This last improvement is needed when creating a KX_VertexProxy
from a KX_PolyProxy, which was impossible before without a
dirty hack.