UPBGE: Fix backface culling for wireframe render.
Previously the wireframe render wasn't using the backface culling setting
from the materials.
To solve this issue a new getter in RAS_IPolyMaterial named IsCullFace is
added to get the backface culling value. After getting this value a call
to SetCullFace is manually done to set the backface culling in case of
wireframe render only.