UPBGE: Move modification flag in RAS_IDisplayArray.
Previously modification flag was per mesh and not per display array
which introduce that if only one vertex of a display array is modified
all the display array or update.

To solve this issue the modification flag is moved only in the
RAS_IDisplayArray.
To achive the transission KX_VertexProxy have to store the display
array instead of the mesh proxy and RAS_MeshObject have to give acces
to the display array from the material index.