UPBGE: Optimize bounding box update.
Previously the bounding box was unique per mesh for a shared mesh and
each object was calling to update the mesh bounding box (if allowed).
Of course the AABB was cached in the mesh, but accessing the mesh calling
a function of check if there's a deformer and then get its AABB, all
of this was costing a lot of instruction.

To reduce the instruction we have to analyze the situation:
The object are only reading a bounding box corresponding to a mesh or
a deformer and never write to it. The deformer and the mesh write and
don't read the bounding box.
To bounding box type had to be taken in account: The mesh bounding box
reading the vertices of all the display arrays and computing an AABB.
The deformer bounding box containing only the AABB which is write by
the deformer.

As there's two types of bounding box, the mesh (RAS_MeshObject) and all
the deformers (subclass of RAS_Deformer) are allocating a bounding box
instance: RAS_BoundingBox for deformer and RAS_MeshBoundingBox only for
mesh.
This instance is set in a bounding box manager (RAS_BoundingBoxManager).
This manager is used to update all the bounding box and owning them.
When the mesh user (RAS_MeshUser) of an object is created, it got a pointer
of the bounding box from the mesh or the deformer. This pointer is used
to access the AABB of the bounding box in KX_GameObject::UpdateBounds
without know and care what is modifying it: mesh or deformer.

The bounding box manager describe bellow, is using to updating the bounding
boxes by calling the virtual function RAS_BoundingBox::Update(force).
This function does nothing for RAS_BoundingBox because it is used as
it only by deformers which set the AABB through RAS_BoundingBox::SetAabb(…).
But for RAS_MeshBoundingBox, Update has for effect to compute an AABB
from the display arrays passed in the bounding box constructor.
In some case a mesh can be unused and it's bounding box will be still
updated. To avoid this case, the bounding box manager use a second
list with only the active bounding box to update. These bounding box
are added in this list thanks to a reference count in RAS_BoundingBox.
This reference is incremented when the bounding box is set in a mesh user
and decremented when removed. Internally, the bounding box have a pointer
to the bounding box manager and add it or remove it to the active bounding
box list when reference count respectively 0 and 1.

Finally the bounding box is owned by a scene and allow to be merged with
an other bounding box manager when a lib load is made.