UPBGE: Add objects LOD option while we render Cube Maps
This was not easy because I had to replicate all the existing LOD system
(except the viewport part) to create a new one because an object can have
LOD for normal render pass computed according to distance from the active
camera and another LOD while we render cubemaps computed according to the
distance from the cubemap rendering camera.

The patch might be simplified and cleaned up. This was just a test to
answer Maujoe feature request. A long copy/paste session.

I don't know if this patch is a good idea because this needs to replace
mesh when we begin cubemap rendering pass and restore mesh after rendering
pass. This might be costfull in terms of performances. I don't know.

But maybe in some cases it could be useful. The debate is opened.