UPBGE: Fix global dict freed in blenderplayer after Py_Finalize.
Previously the globalDict was tried to be free after the call
to Py_Finalize which caused to rasie a segmentation fault due
to that fact that the python commands aren't valid after calling
Py_Finalize.

To solve this issue the globalDict is freed just before destructing
the launcher which finlize python.