UPBGE: Avoid updating deformer display array at conversion.
Previously the RAS_MeshObject was considerated modified
when it was created. This had the side effect to update all
the display array contained in all the deformer using the
the same mesh object. This update was useless as all the
data was just copied in the conversion.

To solve this issue the modified flag is set to 0 as default.