UPBGE: Optimize RAS_IDisplayArray::UpdateFrom.
Previously when the colors or the uvs was needed to be updated
we iterated on all the vertex and after on all the uv or color layers.
But to know the color or uv layers number we called the vertex
function getUvSize or getColorSize for every vertex iteration.

As a display array contains only one type of vertex, the color or
uv layers number is get directly from it instead of the the vertex.
For example if we have 30000 vertices, before we was calling 300000
times getUvSize/getColorSize and now we do it only 2 times.