UPBGE: Fix dulicated vertices.
Since the commit introducing the mutliple color layers we was comparing
all the color layers to know if a vertex is equal to an other. But this
comparaison was always false. At the end all the vertices was unique per
faces even for smoothed meshes.

To solve this issue the color comparaison of char array (which was not
using the alpha channel and a wrong comparaison limit) is replaced by
a comparaison of integer. The integer color value is get through the new
function RAS_ITexVert::getRawRGBA. This same function is reused in
KX_VertexProxy to remove some char array to integer conversions.

Bug reported by youle, Akira_San.