UPBGE: Fix render for invalid filters.
Previously if a filter failed its shader compilation, the render
was incorrect because we trust that all filters are able to transfert
the color to one off screen to an other. This of course in case of
invalid shader is not done and let the off screen color uninitialized.

To solve this issue a basic off screen copy is added in case the
filter shader is invalid.

Bug reported by Akira_San #256.