UPBGE: Implement directional light SSS.
Previously only point lights was able to use the SSS in a Ãmaterial.
Now two other ligths can do it: Sun and Hemi.

To implement these light the SSS shader function is fixed to look like
the implementation found in:
https://machinesdontcare.wordpress.com/2008/10/29/subsurface-scatter-shader/

Also the "attenuation" is multiplied by the visible factor of the light.
THis last operation explain why spot light are still unsupported. Indeed
the visible factor of spot light will create hard edges on the surfaces.