UPBGE: Fix meshes polygons sort.
Previously the polygon sort was always using wrong vertex pointer
because of the inheritance class of RAS_TexVert.

To solve this issue the vertices are accesed from the function
GetVertex instead reading the vertex array. In the same way the
new indices are written with the new function SetIndex in
RAS_IDisplayArray.