UPBGE: Don't sort object for transparency object in shadow render.
Previously the transparent object render for simple shadow was using
a sort. But sorting object for the shadow render is useless because
only alpha clip will have effect on the shadows, and alpha clip is
order independant.

Removing the object sort optimize by two the render time of a transparent
object.