UPBGE: Move thread utils in Common/.
Previously the thread utils was in the Expressions/ part but it was not
related to python. To make the situation more clear a new directory
named Common/ now contains the thread utils. These thread utils are now
prefixed of CM_.

Later the message function will go in the same directory.