UPBGE: Move physics setting from material to object.
Previously the friction, restitution and other physics was
exposed in the material physics panel, but physics settings
are not managed per material and only per object. To avoid
confuse the user on this point, the settings are moved to
the physics panel.

Internally any function related to PHY_MaterialProps are
removed and the properties stored in Material struct are
moved in Object struct.