UPBGE: Implement standardized and colord messages.
Previously all commit used a special message format for error,
warning or debug messages. Nothing was standardized.

To solve this issue macro are added to format message in the
most common cases of error, warning and debug message:
- regular
- python
- python attribut
- python function
- logic brick
- function

The macro are in the same order:
CM_[Warning/Error/Debug/Message](msg)
CM_Python[Warning/Error/Debug](msg)
CM_PythonFunction[Warning/Error/Debug](class, functionname, msg)
CM_PythonAttribut[Warning/Error/Debug](class, attributname, msg)
CM_LogicBrick[Warning/Error/Debug](brick, msg)
CM_Function[Warning/Error/Debug](msg)

For all macro starting by CM_Python*, the message contains the name
of the currently runned script and the current line. These info are
printed like: text.py(1). The script name and the line are in white
bold.
If the macro is for an attribut or a function, the message contains
after the script and line the class name and the attribute name
separated of a dot: KX_GameObject.mass. The total in green without
the dot and the attribut in bold too. If it is a function instead of
an attribut the prefix end with a (...).

For logic brick the message begin by: brickname(objectname), in bold
excepted for the parenthesis.

Finally for a C++ function message, the message begin with:
Class::Function in bold.

All these macro are using commons prefix:
Error: (in red bold)
Warning: (in yellow bold)
Debug: (in bold)