UPBGE: Implement python callbacks for custom shaders.
Previously the user wasn't able to set uniform values per object
using a custom shader and not globally for all the shader.

To solve this lack the easiest way is to give the user the possibility
to use a python callback. This callback is called just before
rendering a mesh with the custom shader.

The callbacks executed receive as arguments the shader to alterate
and the game object to have an identifier.

The callbacks list is accessed under the name of BL_Shader.objectCallback.