UPBGE: Implement python callbacks for custom shaders.
Previously the user wasn't able to set uniform values per object
using a custom shader and not globally for all the objects using
this shader.

To solve this lack, the easiest way is to give the user the possibility
to use a python callback. This callback is called just before
rendering of a mesh with the custom shader.
In the same way, we give the ability to use a callback when the
shader is just bound.

The callbacks list is accessed under the name of BL_Shader.objectCallbacks
and BL_Shader.bindCallbacks.

The callbacks executed receive as arguments the game object to
have an identifier for objectCallbacks and none arguments for
bindCallbacks.

Internally BL_Shader contains an array of python list initialized
in the constructor and accessed through index CALLBACKS_BIND and
CALLBACKS_OBJECT.

The function SetProg is now virtual and overloaded to execute the
callbacks in CALLBACKS_BIND. For CALLBACKS_OBJECT the function
Update(rasty, ms) is added. It execute the callbacks and then call
RAS_Shader::Update(rasty, matrix). This function is added to avoid
using a virtual function with as argument a mesh slot which is invalid
for filters shaders.