UPBGE: Add GPU texture coordinates scrolling
The user was able to set from UI static offset and size for the uv used
in a material texture. But rotation was not alloawed and editing these
value dynamicaly in game was impossible.

To solve the first issue a matrix computation using the value MTex::rot
is added in the shader, the matrix resulting is sent to parallax mapping
method and more generally to mtex_mapping_transform.
In this function the computation order is the following:
rotation, size, offset.

In the same way the offset, size and rotation of the material texture
are allowed to be dynamic under the dynamic uniform enumerations named:
GPU_DYNAMIC_TEX_UVOFFSET
GPU_DYNAMIC_TEX_UVSIZE
GPU_DYNAMIC_TEX_UVROTATION

These dynamics uniform are used if the user uncheck the new texture uv
constant flag: MA_CONSTANT_TEXTURE_UV.

At the end, these value are accessible from python through the pytohn
proxy of BL_Texture under the names:
uvOffset
uvSize
uvRotation