UPBGE: Implement mesh batching.
The user always need to render more meshes with less times, different
categories of optimization could be exploited. One of them is the optimization
of the render of the multiple meshes of the same material which represents
objects. The geometry instancing is helping in this area but it's limitation is
that every objects should use the same mesh too. For unique meshes but sharing
the same material none optimization were possible.

At this point the idea is to make static the objects containing a mesh with a
mesh sharing the same material. If all these objects are static we could merge
the vertex and index array of theirs meshes. This technique is named the mesh
batching.
To achieve this technique we have to use a common list of all the vertex
transformed by the original object transform matrix and a list of indices
corresponding the original sub lists of indices but shifted of the number of
vertices there was before the merge of a part.
The render will then take the both lists like a regular meshes, but the issue
here is that even culled meshes will be rendered. To avoid render these meshes
we do a second list of indices pointing to the beginning of the vertex indices
list per batch part and a list of the indices counts per batch part. So in the
render we pass the correct list of pointer to indices list and count to the
OpenGL function glMultiDrawElements, this function will only render the at the
indices pointer passed with the given counts.

Mesh batching must be able to merge a mesh into the batch of meshes and also
split/restore this mesh.

The python api offering these operations is pretty simple. First the user have
to call the operation from a new class named KX_BatchGroup, this class is
contained in KX_gameObject.batchGroup. KX_BatchGroup exposes the list of
objects merged in the batch thanks to KX_BatchGroup.objects. This class also
expose three functions: merge(), split() and destruct(). The first function
merge the objects passed as argument, the second function split the objects
passed as arguments and finally the destruct function split all the objects and
invalid the KX_BatchGroup instance.
Also KX_BatchGroup need to be created by calling its own constructor with a
valid list of objects, if none merging could be done the constructor delete the
KX_BatchGroup.

The batching group is merging and splitting objects and not only mesh parts of
a material in the user side, so every part of a mesh goes to a batch display
array representing an unique material in the batch group. This batch display
array is the class RAS_BatchDisplayArray which is a sub class of
RAS_DisplayArray with a given vertex template and of RAS_IBatchDisplayArray.
This last class stores all the info of the part merged in the struct Part
contained in the list m_parts, also it defines the function Merge() and Spilt().
The first function takes as argument a display array to copy vertices and
indices from, and a matrix used to transform the vertices. The second function
take only the index of the part to remove.

In Merge() we first registered the new part info into the struct Part, then
every vertices are copied, transformed and added to the display array vertex
list. After this the indices array is add back with a shift corresponding to
the vertex count before the merge. Finally the vertex chache is invalided.
In Split() the indices after the part to remove are moved before in the list
with a negative shift corresponding to the vertex count of the part to remove.
Then the end of the indices array is erased like the vertexes list part.
Finally the Part is removed and the vertex cache is invalided.

So now we know how to merge and split display arrays in a batch display array,
we just have to be able to do so but with a more general case: the mesh user
(=object). This is the goal of the RAS_BatchGroup previously exposed under its
sub class KX_BatchGroup. This class contains pretty similar function than
RAS_BatchDisplayArray: Merge() and Split(). The both functions accepts as
argument the mesh user of the object to merge and for Merge() and a transform
matrix more.
When a merge is proceeded the batch group iterate on all the mesh material part
of the mesh passed indirectly by the mesh user and check if one of these part
match an already created batch display array, if not it create it, then the
function Merge is called with the correct arguments
As the batch group is not owned by one object but by the group of object, we
have to use a reference counting system to delete the batch group when no
anymore mesh user are using it.
The last goal of RAS_BatchGroup is to keep in memory the original display array
of the mesh part merged to copy them back when these part are split.

Notes:
- The cull face setting is now always set in the RAS_MaterialBucket in bind
function instead of in multiple places from KX_BlenderMaterial to
RAS_DisplayArrayBucket.
- Alpha blending is set for geometry instancing and batching before render

The mesh batching give similar time result than geometry instancing but as the
power to accept any kind of meshes of a common material.