UPBGE: Remove display list and vertex array render storage.
This render storage was using very old features which are deprecated since
plentiful versions.

The remove of these wasn't made previously in UPBGE and BFBGE because they was
still faster than new VBO. But in UPBGE some efforts was made to optimize
VBO and let VBO render faster than VA. This inversion was possible thanks to
VAO and bind less VBO.

It's now time to say good bye to vertex array and display list.

With the remove of VA storage, custom storage in UI are useless same as
RAS_IStorage in sources, they had removed so.
The previous fix for font objects which tryed to bind its display array is fixed
by simply forbid create a display array if the mesh and mesh material is to NULL,
which is the case only for font objects.