UPBGE: Optimize VBO modification.
Previously when updating VBO or IBO we was using the glBufferData
function to do allocation and data copy. But in case of copy the
allocation is useless and can cost a lot of time.
To avoid this extra cost, a new function in VBO class is added:
VBO::AllocData. This function allocate the VBO and IBO using
glBufferData function, and functions Update[Data/Indices] use the
function glBufferSubData to avoid reallocate the VBO/IBO.

In the same time the VBO/IBO storage type changed to GL_DYNAMIC_DRAW
the best choice when the VBO/IBO are modified frequently. In the
case the VBO/IBO are not modified the dynamic storage type will
not decrease the render time.