UPBGE: Restore compatibility for material physics settings.
Previously in 0c0d788 the physics settings in material were move in
the object. This is a way more representatif of what UPBGE and Bullet
manage these settings.
But in the move none conversion for compatilitby was implemented.

To solve this issue the variables in Material related to physics settings
are restored and a conversion in versioning_upbge.c is added:
We iterate on all the objects, find the mesh and the first material
using physics. The setting in this first material are copied in the
object.