add mirror option for planar reflections/refractions
In case of reflections/refractions, the render was only nice when
we had shadeless materials. But the problem is that we can't have specular
or light color influence on the mirror. So I use the shader made by A.
Romanov for cubemaps to have the same effect on mirrors.
(mtex_mirror)