UPBGE: Implement implicit cast in CListValue::iterator.
Previously CListValue::iterator was simply std::vector<CValue *>::iterator,
it requested to cast the CValue pointer to an other type in the loop code.
Generally an extra line was dedicated to this operation.

To avoid this line or at least avoir explicit cast, the iterator is now a
cast owning a std::vector::iterator. This class is declared with a template
argulment which is the class name to cast to when deferencing the iterator
for example:

CListValue::iterator<KX_GameObject> it = list->GetBegin()

The class redefines the ++ operator, the * operator for deferencing and the
!= operator, to be used in a loop.

By this way the developer doesn't have to bother of the cast if he set properly
the iterator. Also this class gives less control to the original iterator,
it's safer.