UPBGE: Fix mesh active uv and color layer for unspecified texture.
Previously when a texture wasn't specifying the uv layer or node wasn't
specifying the color layer, the uv and color layer was the first one.

To fix this issue we have to set the uv layer or color layer to the active
uv or color layer in the mesh. We get these values with the function
CustomData_get_active_layer_index. It returns -1 if there's no layers or
an index non-zero based if there's an active layer.

After this we have to store the two active layers, to do this a struct named
RAS_MeshObject::LayersInfo. This struct contains the list of layers used as
bafore and the active color and uv layer.