UPBGE: Fix AABB not initialized to mesh AABB for deformers.
Previously the AABB of a deformer wasn't initialized to the mesh AABB
for the first frame. This caused that animated object could be never
updated because never visible even if the mesh AABB is visible.

To solve this issue a new function is added in RAS_BoundingBox to copy
the AABB from an other bounding box.