UPBGE: Fix blender player usage loop.
Previously when the blenderplayer was executed with wrong arguments
the usage was printed in an infinity loop.

This issue was caused to the use of boolean variable named quit which
was set to true only when the argument were true. So if there was no
file quit was always to false and the usage was printed infinitly.

To solve this issue we avoid updating quit variable but using a function
computing it, this function is named quitGame().