UPBGE: Vibration Actuator improvement
2 new methods (start and stop) have been added to the vibration actuator. The Start method starts the vibration and it can update the parameters of the vibration on
fly too. The Stop method stops the vibration.

Additionally, there is 2 strength parameters instead of one: the "left" one sets the strength of the Low frequency joystick's motor and the "right" one sets the
strength of the High frequency joystick's motor.

Also, to check the vibration capabilities of the controller, 2 new bool attributes have been included (isVibrating and hasVibration).

In the implementation side:

To obtain the vibration, in first place, we checked for the LEFTRIGHT effect. This effect is very well supported in all XInput controllers. In case of this effect
is not supported by the controller we added a back-up SINE effect that it is the most supported in all controllers but XInput ones. The cons of this back-up SINE
effect is that only supports vibration for one of both motors (the low frequency one).
Additionally, we added another effect (CUSTOM). This effect is used by some NOT XInput controllers to make the LEFTRIGHT effect. Also, is the effect most supported
by controllers in the Mac OS.

The vibration algorithm is:

    Define vibration effect type and parameters
    Create (or update if already exists) the effect
    Run the effect.

The vibration duration would have to be done using SDL_HapticGetEffectStatus function but this function is not well supported by the most current controllers. In
its place, we used an own time management. Due to this, we have to process this own time management using the ProcessRumbleStatus function before we can get the
rumble status too.